Move modules in google.api.core.helpers up one level, delete google.api.core.helpers. (#4196)
