+++ core/google/api/core/general_helpers.py
